requirements.txt
asgiref==3.8.1
Django==5.1.7
django-soft-delete==1.0.18
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg-pool==3.2.6
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
